:bug: Arrange specs for the moved resolver properly
