+++ spec/helpers-spec.js
@@ -1,12 +1,9 @@
 'use babel'
 
-import Path from 'path'
 import { it } from './helpers'
 import * as Helpers from '../lib/helpers'
 
 describe('Helpers', function() {
-  const defaultConfig = Helpers.fillConfig({})
-
   describe('fillConfig', function() {
     it('fills empty config', function() {
       const config = Helpers.fillConfig({})
@@ -70,37 +67,6 @@ describe('Helpers', function() {
       const error = Helpers.getError('test')
       expect(error.code).toBe('MODULE_NOT_FOUND')
       expect(error.message).toBe("Cannot find module 'test'")
-    })
-  })
-
-  describe('resolveOnFileSystem', function() {
-    it('rejects with a proper error', async function() {
-      try {
-        await Helpers.resolveOnFileSystem('test', __dirname, defaultConfig)
-        expect(false).toBe(true)
-      } catch (_) {
-        expect(_.code).toBe('MODULE_NOT_FOUND')
-        expect(_.message).toBe("Cannot find module 'test'")
-      }
-    })
-    it('resolves direct files properly', async function() {
-      expect(await Helpers.resolveOnFileSystem('x', __filename, defaultConfig)).toBe(__filename)
-    })
-    it('resolves modules with manifests properly', async function() {
-      const mainFile = Path.normalize(Path.join(__dirname, '..', 'lib', 'index.js'))
-      expect(await Helpers.resolveOnFileSystem('x', Path.dirname(__dirname), defaultConfig)).toBe(mainFile)
-    })
-    it('resolves modules with . or ./ in main properely', async function() {
-      const specDir = Path.join(__dirname, 'fixtures', 'dot-in-main')
-      expect(await Helpers.resolveOnFileSystem('x', specDir, defaultConfig)).toBe(
-        Path.join(specDir, 'index.js')
-      )
-    })
-    it('resolves modules without manifests', async function() {
-      const specDir = Path.join(__dirname, 'fixtures', 'no-manifest')
-      expect(await Helpers.resolveOnFileSystem('x', specDir, defaultConfig)).toBe(
-        Path.join(specDir, 'index.json')
-      )
     })
   })
 })
